Code standardtization pass. Fixed a bunch of warnings.

+++ Nuake/src/Core/OS.h
@@ -7,7 +7,7 @@ namespace Nuake
 	public:
 		static int GetTime() 
 		{
-			return std::chrono::system_clock::now().time_since_epoch().count();
+			return static_cast<int>(std::chrono::system_clock::now().time_since_epoch().count());
 		}
 	};
 }
